Create healpy_smoothing_test_2048.ipynb
This is an iPython Notebook file that shows the simple steps I've tried so far, to smooth the FIS 90 micron map with a 1 degree FWHM Gaussian beam. In order to run it, you will need the HEALPix AKARI 90 micron all-sky mpa (the one at NSIDE resolution 2048, and in NESTED format). You'll also need the healpy, numpy, and astropy python packages, as well as cfitsio.